+++ internal/jwt/jwt.go
@@ -1,4 +1,4 @@
-package crypto
+package jwt
 
 import (
 	"context"
