+++ internal/router/router.go
@@ -37,7 +37,7 @@ func New(handlers *handlers.Handlers, secret string) chi.Router {
 				r.Post("/orders", handlers.Orders.UploadOrder)
 				r.Get("/orders", handlers.Orders.GetOrders)
 				r.Get("/balance", handlers.User.Balance)
-				r.Post("/balance/withdraw", handlers.Withdraw.CreateWithdraw)
+				r.Post("/balance/withdraw", handlers.Withdraw.Withdraw)
 				r.Get("/withdrawals", handlers.Withdraw.GetWithdrawals)
 			})
